refactor(opencode): split sdkCanonicalMapping into canonicalMapping/ subdir
Mirrors the split commit #254 already applied to the claude/codex/acp mappers:
opencode was the lone holdout with a 1105-line monolith. Extracted 10 focused
domain modules under canonicalMapping/ (readers, fileChangeKind, toolClassification,
toolPayload, permissions, questions, usage, subAgents, textItems, dispatch);
sdkCanonicalMapping.ts becomes a 26-line barrel preserving the exact public
surface (state re-exports + mapOpenCodeEvent + closeOpenItems), so sdkSession.ts
and the 35-case sdkCanonicalMapping.test.ts import unchanged.

Pure file move — no logic edits. Verified: 35 mapper tests + 19 sdkSession tests
green, typecheck exit 0, lint 0/0.
